moved fns into loadDOM module
