+++ javascripts/loadDOM.js
@@ -1,0 +1,36 @@
+'use strict';
+
+let $ = require('jquery');
+let getArray = require('./arrayBuilder');
+let userInteraction = require('./userInteractionDOM');
+let templates = require('./templates');
+
+let domLoader = {};
+
+domLoader.attractionsList = function() {
+getArray.attractionsList().then(function(dataFromGetAttractionsList){
+    $('#attractionList').append(templates.testTemplate({list : dataFromGetAttractionsList}));
+		// userInteraction.nameClick();
+});
+};
+
+domLoader.areaList = function() {
+getArray.areas().then(function(dataFromGetAreaList) {
+    let blankGridSpace = {
+        colorTheme: "",
+        decription: "",
+        id: "",
+        name: ""
+    };
+    dataFromGetAreaList.splice(4, 0, blankGridSpace);
+    $('#mapGrid').append( templates.gridTemplate({area: dataFromGetAreaList}) );
+});
+};
+
+domLoader.attractionsList = function() {
+getArray.parkInfo().then(function(dataFromGetParkInfo) {
+    $('#footerDiv').prepend( templates.parkInfo(dataFromGetParkInfo[0]) );
+});
+};
+
+module.exports = domLoader;
